Revert back to onboarding component
