+++ ui/pages/Onboarding/OnboardingImportMetamask.tsx
@@ -7,36 +7,12 @@ import { HIDE_IMPORT_DERIVATION_PATH } from "@tallyho/tally-background/features/
 import SharedButton from "../../components/Shared/SharedButton"
 import SharedBackButton from "../../components/Shared/SharedBackButton"
 import SharedCheckbox from "../../components/Shared/SharedCheckbox"
-import SharedSelect from "../../components/Shared/SharedSelect"
+import OnboardingDerivationPathSelect from "../../components/Onboarding/OnboardingDerivationPathSelect"
 import {
   useBackgroundDispatch,
   useBackgroundSelector,
   useAreKeyringsUnlocked,
 } from "../../hooks"
-
-// TODO make this network specific
-const derivationPaths: { value: string; label: string }[] = [
-  {
-    value: "m/44'/60'/0'/0",
-    label: "ETH (m/44'/60'/0'/0)",
-  },
-  {
-    value: "m/44'/1'/0'/0",
-    label: "ETH Testnet (m/44'/1'/0'/0)",
-  },
-  {
-    value: "m/44'/61'/0'/0",
-    label: "Trezor (m/44'/61'/0'/0)",
-  },
-  {
-    value: "m/44'/137'/0'/0",
-    label: "RSK (m/44'/137'/0'/0)",
-  },
-  {
-    value: "m/44'/37310'/0'/1",
-    label: "RSK Testnet (m/44'/37310'/0'/0)",
-  },
-]
 
 function TextArea({
   value,
@@ -170,12 +146,7 @@ export default function OnboardingImportMetamask(props: Props): ReactElement {
             )}
             {!HIDE_IMPORT_DERIVATION_PATH && isChecked && (
               <div className="select_wrapper">
-                <SharedSelect
-                  placement="top"
-                  placeholder="Select derivation path"
-                  options={derivationPaths}
-                  onChange={setPath}
-                />
+                <OnboardingDerivationPathSelect onChange={setPath} />
               </div>
             )}
           </div>
